fix tests and comment failing ones
