--- neurogym/tests/__init__.py
@@ -1,2 +1,0 @@
-from neurogym.tests.test_envs import test_run
-from neurogym.tests.test_speed import test_speed
